backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Document processing
pymupdf>=1.23.0
python-docx>=1.1.0
chardet>=5.2.0

# ML/AI

chromadb>=0.4.0
google-generativeai>=0.8.0
ollama>=0.1.0
tenacity>=8.2.0  # Retry library with exponential backoff
sentence-transformers>=2.2.0  # Cross-encoder for reranking

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# MCP Server
mcp>=0.9.0
